backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
pandas==2.1.1
python-dotenv==1.0.0
pydantic==2.4.2